+++ src/Configuration.cpp
@@ -30,8 +30,8 @@ Configuration::~Configuration()
 }
 
 int Configuration::base_altitude = 0;
-long double Configuration::base_latitude = 0.0L;
-long double Configuration::base_longitude = 0.0L;
+double Configuration::base_latitude = 0.0;
+double Configuration::base_longitude = 0.0;
 float Configuration::base_geoid = 0.0;
 int Configuration::global_out_port = 0;
 int Configuration::global_ogn_port = 0;
