+++ src/ParserOGN.h
@@ -40,7 +40,7 @@ private:
     // temps
     std::string id;
     int id_t, ac_t, alt, gnd_spd, time, heading;
-    long double lat, lon;
+    double lat, lon;
     float turn_r, climb_r;
 };
