+++ src/Position.cpp
@@ -25,7 +25,7 @@ Position::Position()
 {
 }
 
-Position::Position(long double lat, long double lon, int alt, int time)
+Position::Position(double lat, double lon, int alt, int time)
 : latitude(lat),
   longitude(lon),
   altitude(alt),
@@ -33,7 +33,7 @@ Position::Position(long double lat, long double lon, int alt, int time)
 {
 }
 
-Position::Position(long double lat, long double lon, int alt, int time, float climb_r, float turn_r, int heading)
+Position::Position(double lat, double lon, int alt, int time, float climb_r, float turn_r, int heading)
 : Position(lat, lon, alt, time)
 {
     this->climb_rate = climb_r;
